fixed double naming
