+++ .github/workflows/codequality.yml
@@ -53,7 +53,7 @@ jobs:
         prospector hisim/components/generic_hot_water_storage_modular.py
         prospector hisim/postprocessing/system_chart.py
         prospector hisim/modular_household/component_connections.py
-        prospector examples/modular_household.py
+        prospector examples/modular_example.py
 
 #        # pylint hisim/simulator.py --rcfile=hisim_pylint_config.rc
 #        # flake8 hisim/simulator.py --max-line-length 150
