+++ tests/test_examples.py
@@ -43,9 +43,8 @@ def test_basic_household_with_all_resultfiles():
 
 
 @utils.measure_execution_time
-def test_modular_household_configurations():
-    """ Tests the modular households. """
-    path = "../examples/modular_household.py"
+def test_modular_household_configurations( ):
+    path = "../examples/modular_example.py"
     func = "modular_household_explicit"
     mysimpar = SimulationParameters.one_day_only_with_all_options(year=2019, seconds_per_timestep=60 * 15)
     hisim_main.main(path, func, mysimpar)
